[SYCL][clang] Fix more free-function kernel integration header cases
This commit fixes a number of known issues when the integration header
generates prototypes of the free-function kernels. These changes focus
on the additional removal of aliasing and proper handling of templated
template arguments.

This commit also adds disabled test cases for a known issue with
unresolved nested templated type aliases. These are cases for future
work.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>
Co-authored-by: Sachkov, Alexey <alexey.sachkov@intel.com>