[SYCL] Get rid of device kernel info duplication (#20319)
With the introduction of DeviceKernelInfo, implicit local argument
information is now duplicated in program manager. This patch removes the
duplicate map and makes it so that device kernel info map is filled out
during image registration, with the compile time information added when
it's available (during the first submission of the kernel).

Additinally, this patch ensures that device kernel info instances are
added to the program manager only during image registration, and gets
rid of entries from other sources that polluted the map.