[SYCL] Make `sycl::device` use raw `device_impl *` for its `impl` (#20821)
`device_impl`s are owned by the parent platforms and are only destroyed
when SYCL RT is unloaded. As such, there is no reason to pay the price
of a `std::shared_ptr`, a raw non-owning pointers is enough.

Use a pointer and not a reference because `sycl::device` is assignable.