[SYCL][NFC] Disable warn to error in preprocess test (#20940)
The warn to error setting is not important for this test, as all we want
to do is build the generated preprocessed file. Update the command line
to use -Wno-error to allow for a successful compile. There is a
possibility of certain generated preprocessed file conditions where
warnings will be emitted that we want to not error on.