[clang] Accept non const static work_group_static variables in templates
In general, SYCL doesn't allow non const static variables, but for some
extensions like work_group_static it is allowed. This is done by
checking presence of an attribute on the type of the variable.
However detection of such cases worked incorrectly in template contexts
since a dependent type may not yet have the attribute.
This patch fixes the problem by not diagnosing variables with dependent types.