[SYCL][NFC] Refactor addSYCLDefaultTriple (#20938)
This patch merges two loops traversing SYCL triples into one.
Also optimizes the usage of addSYCLDefaultTriple inside
CreateOffloadingDeviceToolChains. The default triple is guranted to be
added at the beginning of the list. Instead of building the default
triple again in CreateOffloadingDeviceToolChains, we re-use the value
added by addSYCLDefaultTriple to the list of the triples.