[SYCL][ABI-break] Simplify `interop_handle` constructor (#20965)
According to spec "The interop_handle class is an abstraction over the
queue which is being used to invoke the host task and its associated
device and context." and currently interop_handle object is created
internally at the time of submission to the queue, so queue is always
supposed to be available. Device and context are the ones associated
with the queue, so it seems that those parameters in the constructor are
redundant.
We can't get rid of shared_ptr to the queue because interop_handle has
copy/move constructors, so theoretically it can be copied/moved outside
of the scope of the queue object.